+++ packages/components-js/projects/partials/tests/unit/specs/getInitialStyles.spec.ts
@@ -52,16 +52,16 @@ describe('format: jsx', () => {
 describe('format: sha256', () => {
   it('should return hash for core styles', async () => {
     const result = getInitialStyles({ format: 'sha256' });
-    expect(result).toMatchInlineSnapshot(`"'sha256-ZfnOhvEX2eOryDFV5e2OfWReSZ3vdOp/+4BVMriCr8s='"`);
+    expect(result).toMatchInlineSnapshot(`"'sha256-XBDzTbtSMmUmNgWQq4IZiPE9vZ3tUfihakhHws/xW7g='"`);
   });
 
   it('should return hash for core styles with custom prefix', async () => {
     const result = getInitialStyles({ format: 'sha256', prefix: 'custom-prefix' });
-    expect(result).toMatchInlineSnapshot(`"'sha256-RXIuVwqtHEAGJHEmCemmf8YFKX3cwaHXViAbeKr0ZW4='"`);
+    expect(result).toMatchInlineSnapshot(`"'sha256-o6AyOFxaQ3E++499cFOp5eEW6HX2AK9ZhiZyTLEKO6w='"`);
   });
 
   it('should return hash for core styles with multiple custom prefixes', async () => {
     const result = getInitialStyles({ format: 'sha256', prefix: ['', 'some-prefix', 'another-prefix'] });
-    expect(result).toMatchInlineSnapshot(`"'sha256-JuCtI7XAfve+WrlX2UfjLoiaB0GMYPNzoSjQu22owp8='"`);
+    expect(result).toMatchInlineSnapshot(`"'sha256-X8fDsvbBAXDPNcdOUTAJXO+fDFZetx9v/Aeus5GzFlI='"`);
   });
 });